+++ src/index.ts
@@ -2,5 +2,7 @@ export * from "./model.js";
 export * from "./boundary.js";
 export * from "./hierarchy.js";
 export * from "./normalize.js";
+export * from "./quotient.js";
 export * from "./semantic.js";
+export * from "./validate.js";
 export * from "./graph.js";